bots/requirements.txt
# Trading Bot Python Dependencies

# WebSocket
websocket-client>=1.6.0

# Binance API
python-binance>=1.0.19

# Machine Learning
lightgbm>=4.0.0
pandas>=2.0.0
numpy>=1.24.0

# HTTP Requests
requests>=2.31.0

# Optional: For advanced features
# scikit-learn>=1.3.0
